Slight rebalances, fix plant health anim
- Failing to parse a behavior now fails hard instead of silently ignoring

+++ src/main/java/com/lovetropics/minigames/common/content/biodiversity_blitz/plot/plant/state/PlantHealth.java
@@ -23,7 +23,7 @@ public final class PlantHealth {
 		return this.health;
 	}
 
-	public float healthPercent() {
-		return this.health / (float) this.maxHealth;
+	public double healthPercent() {
+		return this.health / (double) this.maxHealth;
 	}
 }
